refactor: Relocate extensions.ts in Editor helpers
Moved the extensions.ts file from the Editor folder to the Editor/helpers subfolder. This change also includes updating related import paths in useEditorCoreInit.ts and useEditorFileInit.ts to reflect this new location of the extensions file. This reorganization improves the structure and readability of the code.
